Replace pint with astropy.units
This will increase the size of our requirements, but I think it is
worthy. I was comparing it with pint and found that it works better
sometimes. But if we have problems in the future with it, it is not
hard to go back to pint, they have a similar api, so we will be able to
replace it with little work I think.